third_party/WebKit/LayoutTests/virtual/stable/http/tests/sendbeacon/beacon-cross-origin-redirect-blob-expected.txt

PingLoader dispatched to 'http://127.0.0.1:8080/resources/redirection-response.php?status=302&simple=true&target=/non-existent.php'.
PingLoader dispatched to 'http://127.0.0.1:8080/resources/redirection-response.php?status=302&simple=false&target=/non-existent.php'.
CONSOLE ERROR: Redirect from 'http://127.0.0.1:8080/resources/redirection-response.php?status=302&simple=false&target=/non-existent.php' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Verifying navigator.sendBeacon(Blob) non-CORS cross-origin redirect handling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.sendBeacon("http://127.0.0.1:8080/resources/redirection-response.php?status=302&simple=true&target=/non-existent.php", blob); is true
PASS navigator.sendBeacon("http://127.0.0.1:8080/resources/redirection-response.php?status=302&simple=false&target=/non-existent.php", blob); is true
PASS successfullyParsed is true

TEST COMPLETE
